source/requirements.txt
pymupdf
faiss-gpu
transformers
torch
ollama
chromadb
scikit-learn
nltk
streamlit~=1.37.0
tinydb
langchain~=0.3.0
langchain_community
langchain_ollama
langchain_chroma
langchain_text_splitters
langchain_core
langchain_openai
python-dotenv
pydantic
jupyter~=1.1.1
pip~=24.2
emoji~=2.12.1
attrs~=23.2.0
build~=1.2.1
toml~=0.10.2
wheel~=0.41.2
lxml~=5.3.0
pillow~=10.3.0
tornado~=6.4.1
Jinja2~=3.1.4
setuptools~=68.2.0
typing_extensions~=4.12.0
langgraph~=0.2.26
pyarrow
streamlit-authenticator~=0.4.1